Added drawing fucntions and rearanged dependencies
