+++ drawing.c
@@ -1,0 +1,15 @@
+#include "drawinghead.c"
+
+// Functions used for drawing
+
+drawCircle(struct DrawingHead *head) {
+
+}
+
+drawSquare(struct DrawingHead *head) {
+
+}
+
+drawTriangle(struct DrawingHead *head) {
+
+}
